backend/requirements.txt
flask_sqlalchemy
mysql-connector-python
python-dotenv
bcrypt
